[#53] refactor: rating float -> int

+++ src/main/java/com/napzak/domain/review/api/dto/response/StoreReviewDto.java
@@ -6,7 +6,7 @@ import com.napzak.domain.review.core.vo.Review;
 public record StoreReviewDto(
 	Long reviewId,
 	String reviewerNickname,
-	float rating,
+	int rating,
 	String comment,
 	Long relatedProductId,
 	String relatedProductName
